Move hash() function
